refactor: move test-support module to spark and remove
