+++ spark/src/test/java/org/neo4j/spark/DataSourceReaderTypesTSE.java
@@ -21,6 +21,8 @@ import org.apache.spark.sql.Row;
 import org.apache.spark.sql.catalyst.expressions.GenericRowWithSchema;
 import org.junit.Test;
 import org.neo4j.driver.Session;
+import org.neo4j.spark.testsupport.SparkConnectorScalaBaseTSE;
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteIT;
 
 import java.sql.Timestamp;
 import java.time.*;
