+++ spark/src/test/java/org/neo4j/spark/SparkConnectorSuiteIT.java
@@ -18,6 +18,7 @@ package org.neo4j.spark;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteIT;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
