+++ spark/src/test/java/org/neo4j/spark/testsupport/Assert.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.spark;//
+package org.neo4j.spark.testsupport;//
 // Source code recreated from a .class file by IntelliJ IDEA
 // (powered by Fernflower decompiler)
 //
