+++ spark/src/test/resources/simplelogger.properties
