+++ spark/src/test/scala/org/neo4j/spark/CommonTestSuiteIT.scala
@@ -19,6 +19,7 @@ package org.neo4j.spark
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
 import org.neo4j.spark.service.SchemaServiceTSE
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteIT
 
 @RunWith(classOf[Suite])
 @Suite.SuiteClasses(Array(
