+++ spark/src/test/scala/org/neo4j/spark/CommonTestSuiteWithApocIT.scala
@@ -19,6 +19,7 @@ package org.neo4j.spark
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
 import org.neo4j.spark.service.SchemaServiceWithApocTSE
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteWithApocIT
 
 @RunWith(classOf[Suite])
 @Suite.SuiteClasses(Array(
