+++ spark/src/test/scala/org/neo4j/spark/DataSourceAggregationTSE.scala
@@ -19,6 +19,8 @@ package org.neo4j.spark
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.neo4j.driver.TransactionContext
+import org.neo4j.spark.testsupport.SparkConnectorScalaBaseTSE
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteIT
 
 class DataSourceReaderAggregationTSE extends SparkConnectorScalaBaseTSE {
