+++ spark/src/test/scala/org/neo4j/spark/DataSourceReaderWithApocTSE.scala
@@ -24,6 +24,8 @@ import org.neo4j.driver.Transaction
 import org.neo4j.driver.TransactionCallback
 import org.neo4j.driver.TransactionContext
 import org.neo4j.driver.summary.ResultSummary
+import org.neo4j.spark.testsupport.SparkConnectorScalaBaseWithApocTSE
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteWithApocIT
 
 import java.sql.Timestamp
 import java.time.Instant
