+++ spark/src/test/scala/org/neo4j/spark/SparkConnector30ScalaSuiteWithApocIT.scala
@@ -18,6 +18,7 @@ package org.neo4j.spark
 
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteWithApocIT
 
 @RunWith(classOf[Suite])
 @Suite.SuiteClasses(Array(
