+++ spark/src/test/scala/org/neo4j/spark/service/SchemaServiceWithApocTSE.scala
@@ -24,15 +24,15 @@ import org.junit.Before
 import org.junit.FixMethodOrder
 import org.junit.Test
 import org.junit.runners.MethodSorters
-import org.neo4j.Closeables.use
 import org.neo4j.caniuse.Neo4j
 import org.neo4j.caniuse.Neo4jDeploymentType
 import org.neo4j.caniuse.Neo4jEdition
 import org.neo4j.caniuse.Neo4jVersion
 import org.neo4j.driver.TransactionContext
-import org.neo4j.spark.SparkConnectorScalaBaseWithApocTSE
-import org.neo4j.spark.SparkConnectorScalaSuiteWithApocIT
 import org.neo4j.spark.converter.CypherToSparkTypeConverter
+import org.neo4j.spark.testsupport.Closeables.use
+import org.neo4j.spark.testsupport.SparkConnectorScalaBaseWithApocTSE
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteWithApocIT
 import org.neo4j.spark.util.DriverCache
 import org.neo4j.spark.util.Neo4jOptions
 import org.neo4j.spark.util.Neo4jUtil
