+++ spark/src/test/scala/org/neo4j/spark/testsupport/Closeables.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j
+package org.neo4j.spark.testsupport
 
 object Closeables {
