+++ spark/src/test/scala/org/neo4j/spark/util/Neo4jOptionsIT.scala
@@ -20,9 +20,9 @@ import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Ignore
 import org.junit.Test
-import org.neo4j.Closeables.use
-import org.neo4j.spark.SparkConnectorScalaSuiteIT
-import org.neo4j.spark.SparkConnectorScalaSuiteIT.server
+import org.neo4j.spark.testsupport.Closeables.use
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteIT
+import org.neo4j.spark.testsupport.SparkConnectorScalaSuiteIT.server
 
 class Neo4jOptionsIT extends SparkConnectorScalaSuiteIT {
